updated road generation
Improved road generation, it now reads from an export from dfhack, the
lua script for that is located in samples/dorfgen.zip.  the script is to
be run after the world is loaded, so either in the pre-embark screen of
fortress mode, or in legends mode.

replaced uggrass with roadsurface gravel, which replaced snow on top
with air.  it normally drops vanilla gravel.