added site terrain loading from str image
Added the loading of types of land around sites from the str image.  it
tries to place the villages on the areas in the str image which should
have buildings on it, it currently doesn't check for the correct tile
for lairs or hippyhuts though, so they are currently not spawning.

fixed lakes not working, will require deleting the biomes.csv in config
for it to work.